cpp: FormatReader detail: Add openBytes implementation method
As for isThisType, the intent is that openBytesImpl should be
implemented by reader implementations, while the FormatReader
implementation implements variants of openBytes.  This is to
remove the need for all readers to implement all openBytes
variants due to the need implement all overloaded virtual
methods in derived classes to avoid methods being masked.