cmake: Add model-source-only option
This is a temporary change to permit the model source to be
generated without a dependency upon ome-common, which isn't
needed for source generation.  The source will not build
with this option enabled.

When the model code is split into a separate repository, this
option can be moved there, or redone in a better way.