cpp: bioformats: Ignore fileOnly on close
While this is used in the reader interface, which is shared
with the writer interface, it isn't used in the Java writers,
and can lead to double-close attempts if called explicitly
and then again in the destructor (since currentId isn't
invalidated).  This can lead to writing a corrupt IFD in the
case of TIFF formats, since nextIFD() can be called twice
without the IFD being set up properly.