OMETIFFReader: Reduce the number of times OME-XML metadata is parsed
Cache the OME-XML metadata and filename for isThisType, isSingleFile
and initFile methods.  This metadata will avoid repeated expensive
reparsing of the metadata if used multiple times with the same file.
Note: not thread safe.

Default isThisType check will now fail if the file being checked
does not exist.  This is because we canonicalise the path to ensure
the cached path will always correctly compare whatever path is
used to any of the methods.  It also ensures true will only be
returned for files which both exist and which are valid.

Update isThisType unit tests to check using real files as well as
dummy data.