OMETIFFReader: initFile: Factor out image count fixup
This check is a little odd.  Firstly the problem should be
logically impossible to occur in practice, since the
image counts are always computed directly from the metadata
values.  Secondly, it's only performed for the first series.
Lastly, it is performed unconditionally and unnecessarily
if the image count is 1.

Restrict the "correction" to occur for all series, but only
if the metadata is truly incorrect.  If any correction is
done, log a warning so that there is some indication that
the metadata required manual correction.