FormatWriter: Expose basic metadata query methods
Present in the implementation detail as a convenience for writer
implmentations, this exposes the same information for the
convenience of the end user.  This is useful to introspect the
basic image dimensionality as it can vary with sub-resolution
level.  It can also be useful for testing the implementation
to check that the writer's expectations are matching the caller's
and the state in the MetadataRetrieve instance.