PixelBuffer: Decrease size from 9D to 4D
Only 3 dimensions are used in practice (X, Y, S).  The others
are not used by any openBytes, saveBytes or other operations,
but the per-pixel cost of computing indexes into the pixel
buffer are high.  This improves performance by over a factor of
two (9/5 = 2.25× improvement), which will greatly improve the
speed of many common operations, from reading and writing
pixel data, to copying data in and out of the pixel buffer.

This also greatly simplifies the API and makes it much easier
to use.  There are now only two types of pixel buffer:
  - interleaved (SXYZ) "chunky"
  - non-interleaved (XYZS) "planar"
so an end user will no longer have to contend with pixel buffer
ordering mismatches because the DimensionOrder differs subtly;
it's no longer possible to have that type of mismatch, because
DimensionOrder is no longer coupled to the pixel buffer order.

Lastly, for uniformity with existing practice and to make the
API a bit more coherent, "subchannels" are renamed to "samples"
to make the terminology compatible with libtiff and other
imaging libraries.